LUN-328: renamed fkt name, moved type, resolved issues
