+++ src/constants/labels.json
@@ -11,7 +11,7 @@
     "progressDescription": "Modulen erledigt",
     "start": "Beginnen",
     "continue": "Fortsetzen",
-    "checkModules": "Module ansehen",
+    "viewModules": "Module ansehen",
     "addCustomDiscipline": "Bereich hinzufügen",
     "customDisciplineSection": "Individueller Lernbereich",
     "customDisciplineExplanation": "Hier kannst du eigene Lernbereiche hinzufügen",
